+++ rollup.config.js
@@ -4,9 +4,13 @@ import pkg from './package.json'
 export default {
     input: 'src/index.js',
     output: {
+        file: 'c3.js',
         name: 'c3',
         format: 'umd',
-        banner: `/* @license C3.js v${pkg.version} | (c) C3 Team and other contributors | http://c3js.org/ */`
+        banner: `/* @license C3.js v${pkg.version} | (c) C3 Team and other contributors | http://c3js.org/ */`,
+        globals:{
+            d3:'d3'
+        }
     },
     plugins: [babel({
         presets: [['es2015', {
@@ -15,5 +19,6 @@ export default {
         plugins: [
             'external-helpers'
         ]
-    })]
+    })],
+    external: ['d3'],
 };
